+++ src/components/deck/deck.tsx
@@ -37,6 +37,7 @@ import defaultTheme, {
 import { defaultTransition, SlideTransition } from '../transitions';
 import { SwipeEventData } from 'react-swipeable';
 import { MarkdownComponentMap } from '../../utils/mdx-component-mapper';
+import TemplateWrapper from '../template-wrapper';
 
 export type DeckContextType = {
   deckId: string | number;
@@ -75,6 +76,7 @@ export type DeckContextType = {
 };
 
 export const DeckContext = createContext<DeckContextType>(null as any);
+DeckContext.displayName = 'DeckContext';
 const noop = () => {};
 
 /**
@@ -177,8 +179,12 @@ export const DeckInternal = forwardRef<DeckRef, DeckInternalProps>(
       cancelTransition
     } = useDeckState(initialDeckState);
 
-    const [setPlaceholderContainer, slideIds, slideIdsInitialized] =
-      useCollectSlides();
+    const [
+      setPlaceholderContainer,
+      slideIds,
+      slideIdsWithTemplates,
+      slideIdsInitialized
+    ] = useCollectSlides();
 
     // It really is much easier to just expose methods to the outside world that
     // drive the presentation through its state rather than trying to implement a
@@ -377,6 +383,17 @@ export const DeckInternal = forwardRef<DeckRef, DeckInternalProps>(
       backdropStyle['backgroundColor'] = 'black';
     }
 
+    const doesCurrentSlideHaveItsOwnTemplate =
+      slideIdsWithTemplates.includes(activeSlideId);
+
+    const templateElement: ReactNode =
+      typeof template === 'function'
+        ? template?.({
+            slideNumber: activeView.slideIndex + 1,
+            numberOfSlides: slideIds.length
+          })
+        : template;
+
     return (
       <ThemeProvider
         theme={mergeTheme({
@@ -392,12 +409,6 @@ export const DeckInternal = forwardRef<DeckRef, DeckInternalProps>(
             overflow: 'hidden'
           }}
         >
-          <Portal
-            ref={setSlidePortalNode}
-            overviewMode={overviewMode}
-            printMode={printMode}
-            fitAspectRatioStyle={fitAspectRatioStyle}
-          />
           <DeckContext.Provider
             value={{
               deckId,
@@ -435,6 +446,26 @@ export const DeckInternal = forwardRef<DeckRef, DeckInternalProps>(
               backgroundImage
             }}
           >
+            <Portal
+              ref={setSlidePortalNode}
+              overviewMode={overviewMode}
+              printMode={printMode}
+              fitAspectRatioStyle={fitAspectRatioStyle}
+            >
+              {!doesCurrentSlideHaveItsOwnTemplate && !overviewMode && (
+                <TemplateWrapper
+                  style={{
+                    ...wrapperStyle,
+                    // Slides are appended to the parent as they are portaled in and end up later in
+                    // the source order. Adding zIndex to the template to overlay the sibling slides
+                    // once they have been portaled in.
+                    zIndex: 1
+                  }}
+                >
+                  {templateElement}
+                </TemplateWrapper>
+              )}
+            </Portal>
             <div ref={setPlaceholderContainer} style={{ display: 'none' }}>
               {children}
             </div>