+++ src/components/slide/slide.tsx
@@ -24,6 +24,7 @@ import { ActivationThresholds, useCollectSteps } from '../../hooks/use-steps';
 import { GOTO_FINAL_STEP } from '../../hooks/use-deck-state';
 import { useSwipeable } from 'react-swipeable';
 import { SlideTransition } from '../transitions';
+import TemplateWrapper from '../template-wrapper';
 
 const noop = () => {};
 
@@ -36,6 +37,7 @@ export type SlideContextType = {
 };
 
 export const SlideContext = createContext<SlideContextType>(null as any);
+SlideContext.displayName = 'SlideContext';
 
 type SlideContainerProps = BackgroundProps &
   ColorProps & { backgroundOpacity: number };
@@ -73,15 +75,6 @@ const SlideWrapper = styled.div<ColorProps & SpaceProps>(
   `
 );
 
-const TemplateWrapper = styled.div`
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  pointer-events: none;
-`;
-
 export const AnimatedDiv = styled(animated.div)`
   width: 100%;
   height: 100%;
@@ -115,7 +108,8 @@ const Slide = (props: SlideProps): JSX.Element => {
     throw new Error(`Slide components may not be nested within each other.`);
   }
 
-  const { slideId, placeholder } = useSlide(userProvidedId);
+  const hasTemplate = Boolean(template);
+  const { slideId, placeholder } = useSlide(hasTemplate, userProvidedId);
   const { setStepContainer, activationThresholds, finalStepIndex } =
     useCollectSteps();
   const {
@@ -312,20 +306,24 @@ const Slide = (props: SlideProps): JSX.Element => {
     };
   }, [wrapperOverrideStyle, theme, padding]);
 
-  const templateFn =
-    typeof template === 'function'
-      ? template
-      : typeof deckTemplate === 'function'
-      ? deckTemplate
-      : null;
+  const slideTemplateElement: ReactNode = useMemo(() => {
+    return typeof template === 'function'
+      ? template?.({
+          slideNumber: activeView.slideIndex + 1,
+          numberOfSlides: slideCount
+        })
+      : template;
+  }, [template, activeView.slideIndex, slideCount]);
 
-  const templateElement: ReactNode =
-    templateFn?.({
-      slideNumber: activeView.slideIndex + 1,
-      numberOfSlides: slideCount
-    }) ||
-    template ||
-    deckTemplate;
+  const deckTemplateElement: ReactNode = useMemo(() => {
+    if (!inOverviewMode) return null;
+    return typeof deckTemplate === 'function'
+      ? deckTemplate?.({
+          slideNumber: activeView.slideIndex + 1,
+          numberOfSlides: slideCount
+        })
+      : deckTemplate;
+  }, [inOverviewMode, deckTemplate, activeView.slideIndex, slideCount]);
 
   const swipeHandler = useSwipeable({
     onSwiped: (eventData) => onMobileSlide(eventData)
@@ -370,9 +368,16 @@ const Slide = (props: SlideProps): JSX.Element => {
                 color={textColor}
                 {...swipeHandler}
               >
-                <TemplateWrapper style={wrapperOverrideStyle}>
-                  {templateElement}
-                </TemplateWrapper>
+                {isActive && hasTemplate && (
+                  <TemplateWrapper style={wrapperOverrideStyle}>
+                    {slideTemplateElement}
+                  </TemplateWrapper>
+                )}
+                {inOverviewMode && (
+                  <TemplateWrapper style={wrapperOverrideStyle}>
+                    {deckTemplateElement}
+                  </TemplateWrapper>
+                )}
                 <SlideWrapper
                   style={scaledWrapperOverrideStyle}
                   padding={padding}