+++ src/hooks/use-slides.tsx
@@ -12,6 +12,8 @@ export function useCollectSlides() {
   const [initialized, setInitialized] = useState(false);
   const [slideContainer, setSlideContainer] = useState<HTMLElement | null>();
   const [slideIds, setSlideIds] = useState<SlideId[]>([]);
+  const [slideIdsOfSlidesWithTemplates, setSlideIdsOfSlidesWithTemplates] =
+    useState<SlideId[]>([]);
 
   useEffect(() => {
     if (!slideContainer) return;
@@ -20,23 +22,40 @@ export function useCollectSlides() {
     ) as unknown as Iterable<HTMLElement>;
 
     const nextSlideIds: SlideId[] = [];
+    const nextSlideIdsOfSlidesWithTemplates: SlideId[] = [];
     for (const placeholderNode of slides) {
-      const { slideId } = placeholderNode.dataset;
+      const { slideId, slideHasTemplate } = placeholderNode.dataset;
       nextSlideIds.push(slideId!);
+      if (slideHasTemplate === 'true') {
+        nextSlideIdsOfSlidesWithTemplates.push(slideId!);
+      }
     }
     setSlideIds(nextSlideIds);
+    setSlideIdsOfSlidesWithTemplates(nextSlideIdsOfSlidesWithTemplates);
     setInitialized(true);
   }, [slideContainer]);
 
-  return [setSlideContainer, slideIds, initialized] as const;
+  return [
+    setSlideContainer,
+    slideIds,
+    slideIdsOfSlidesWithTemplates,
+    initialized
+  ] as const;
 }
 
-export function useSlide(userProvidedId?: SlideId) {
+export function useSlide(
+  doesSlideHaveTemplate: boolean,
+  userProvidedId?: SlideId
+) {
   const [slideId] = useState<SlideId>(userProvidedId || ulid);
   return {
     slideId,
     placeholder: (
-      <div className={PLACEHOLDER_CLASS_NAME} data-slide-id={slideId} />
+      <div
+        className={PLACEHOLDER_CLASS_NAME}
+        data-slide-id={slideId}
+        data-slide-has-template={doesSlideHaveTemplate}
+      />
     )
   };
 }